+++ src/main/java/xyz/carnage/entity/EntitiesRegistry.java
@@ -1,13 +1,12 @@
 package xyz.carnage.entity;
 
-import net.fabricmc.fabric.api.object.builder.v1.entity.FabricEntityTypeBuilder;
-import net.minecraft.entity.EntityDimensions;
 import net.minecraft.entity.EntityType;
 import net.minecraft.entity.SpawnGroup;
 import net.minecraft.registry.Registries;
 import net.minecraft.registry.Registry;
 import net.minecraft.util.Identifier;
 import xyz.carnage.Carnage;
+import xyz.carnage.entity.customEntities.BrineBreakerEntity;
 
 public class EntitiesRegistry {
     public static final EntityType<BrineBreakerEntity> BRINEBREAKER = Registry.register(Registries.ENTITY_TYPE,
