+++ src/main/java/xyz/carnage/entity/client/entityModels/BrinebreakerEntityModel.java
@@ -1,16 +1,13 @@
-package xyz.carnage.entity.client;
+package xyz.carnage.entity.client.entityModels;
 
 import net.minecraft.client.model.*;
 import net.minecraft.client.render.VertexConsumer;
 import net.minecraft.client.render.entity.model.EntityModel;
 import net.minecraft.client.render.entity.model.EntityModelLayer;
 import net.minecraft.client.util.math.MatrixStack;
-import net.minecraft.entity.Entity;
-import net.minecraft.sound.SoundEvent;
-import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.Identifier;
 import xyz.carnage.Carnage;
-import xyz.carnage.entity.BrineBreakerEntity;
+import xyz.carnage.entity.customEntities.BrineBreakerEntity;
 
 // Made with Blockbench 4.12.2
 // Exported for Minecraft version 1.17+ for Yarn
