+++ src/main/java/xyz/carnage/itemmgmt/items/BrinebreakerItem.java
@@ -9,7 +9,7 @@ import net.minecraft.world.World;
 import xyz.carnage.Carnage;
 import xyz.carnage.combos.ComboManager;
 import xyz.carnage.combos.ComboTracker;
-import xyz.carnage.entity.BrineBreakerEntity;
+import xyz.carnage.entity.customEntities.BrineBreakerEntity;
 import xyz.carnage.entity.EntitiesRegistry;
 
 public class BrinebreakerItem extends TridentItem {
